Introduce `ClassValue`
