+++ base/src/main/java/io/spine/type/ClassValue.java
@@ -1,0 +1,80 @@
+/*
+ * Copyright 2017, TeamDev Ltd. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.type;
+
+import javax.annotation.Nullable;
+import java.util.Objects;
+
+/**
+ * Abstract base for classes holding a value of a {@link Class}.
+ *
+ * @param <T> the type of the class
+ *
+ * @author Alexander Yevsyukov
+ */
+public abstract class ClassValue<T> {
+
+    private final Class<? extends T> value;
+
+    public ClassValue(Class<? extends T> value) {
+        this.value = value;
+    }
+
+    /** Returns value of the object. */
+    public Class<? extends T> value() {
+        return this.value;
+    }
+
+    /**
+     * Returns {@linkplain Class#getName() the name} of the enclosed class value.
+     *
+     * @return the value class name
+     */
+    @Override
+    public String toString() {
+        return value.getName();
+    }
+
+    /** Obtains the {@code ClassName} for this message class. */
+    public ClassName getClassName() {
+        return ClassName.of(value());
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public int hashCode() {
+        return Objects.hash(value);
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public boolean equals(@Nullable Object obj) {
+        if (this == obj) {
+            return true;
+        }
+
+        if (obj == null || getClass() != obj.getClass()) {
+            return false;
+        }
+        final ClassValue other = (ClassValue) obj;
+        return Objects.equals(this.value, other.value);
+    }
+}
