+++ base/src/main/java/io/spine/type/MessageClass.java
@@ -32,52 +32,10 @@ import static com.google.common.base.Preconditions.checkNotNull;
  *
  * @author Alexander Yevsyukov
  */
-public abstract class MessageClass {
-
-    private final Class<? extends Message> value;
+public abstract class MessageClass extends ClassValue<Message> {
 
     protected MessageClass(Class<? extends Message> value) {
+        super(value);
         checkNotNull(value);
-        this.value = value;
-    }
-
-    /** Returns value of the object. */
-    public Class<? extends Message> value() {
-        return this.value;
-    }
-
-    /**
-     * Returns {@linkplain Class#getName() the name} of the enclosed class value.
-     *
-     * @return the value class name
-     */
-    @Override
-    public String toString() {
-        return value.getName();
-    }
-
-    /** Obtains the {@code ClassName} for this message class. */
-    public ClassName getClassName() {
-        return ClassName.of(value());
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    public int hashCode() {
-        return Objects.hash(value);
-    }
-
-    /** {@inheritDoc} */
-    @Override
-    public boolean equals(@Nullable Object obj) {
-        if (this == obj) {
-            return true;
-        }
-
-        if (obj == null || getClass() != obj.getClass()) {
-            return false;
-        }
-        final MessageClass other = (MessageClass) obj;
-        return Objects.equals(this.value, other.value);
     }
 }
